rename cm-nspell to cm-espells
