+++ client/modules/.pnpm-debug.log
@@ -1,0 +1,20 @@
+{
+  "0 debug pnpm:scope": {
+    "selected": 1,
+    "workspacePrefix": "D:\\programming\\wikijump\\client"
+  },
+  "1 error pnpm": {
+    "errno": 1,
+    "code": "ELIFECYCLE",
+    "pkgid": "root@",
+    "stage": "validate",
+    "script": "ultra --concurrency 10 validate:cmd",
+    "pkgname": "root",
+    "err": {
+      "name": "Error",
+      "message": "root@ validate: `ultra --concurrency 10 validate:cmd`\nExit status 1",
+      "code": "ELIFECYCLE",
+      "stack": "Error: root@ validate: `ultra --concurrency 10 validate:cmd`\nExit status 1\n    at EventEmitter.<anonymous> (C:\\Users\\Monkatraz\\AppData\\Roaming\\npm\\pnpm-global\\5\\node_modules\\.pnpm\\pnpm@6.7.1\\node_modules\\pnpm\\dist\\pnpm.cjs:103960:20)\n    at EventEmitter.emit (node:events:365:28)\n    at ChildProcess.<anonymous> (C:\\Users\\Monkatraz\\AppData\\Roaming\\npm\\pnpm-global\\5\\node_modules\\.pnpm\\pnpm@6.7.1\\node_modules\\pnpm\\dist\\pnpm.cjs:92266:18)\n    at ChildProcess.emit (node:events:365:28)\n    at maybeClose (node:internal/child_process:1067:16)\n    at Process.ChildProcess._handle.onexit (node:internal/child_process:301:5)"
+    }
+  }
+}
