+++ client/modules/cm-espells/src/dicts.ts
