+++ client/modules/cm-espells/src/espells/nspell.worker.ts
