+++ client/modules/cm-espells/src/hover.ts
