+++ client/modules/cm-espells/src/locales.ts
