+++ client/modules/cm-espells/src/tokenizer.ts
